Set triggerv2 as default controller and dispatcher for KEDA scaling
