+++ control-plane/pkg/reconciler/trigger/trigger_test.go
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Knative Authors
+ * Copyright 2021 The Knative Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,63 +20,55 @@ import (
 	"context"
 	"fmt"
 	"testing"
-	"time"
 
-	"github.com/Shopify/sarama"
-	"github.com/google/go-cmp/cmp"
-	"google.golang.org/protobuf/testing/protocmp"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
-	clientgotesting "k8s.io/client-go/testing"
 	"k8s.io/utils/pointer"
 	sources "knative.dev/eventing-kafka/pkg/apis/sources/v1beta1"
-	"knative.dev/pkg/apis"
-	duckv1 "knative.dev/pkg/apis/duck/v1"
-	kubeclient "knative.dev/pkg/client/injection/kube/client/fake"
-	"knative.dev/pkg/controller"
-	"knative.dev/pkg/logging"
-	. "knative.dev/pkg/reconciler/testing"
-	"knative.dev/pkg/resolver"
-	"knative.dev/pkg/tracker"
-
 	eventingduck "knative.dev/eventing/pkg/apis/duck/v1"
+
+	internals "knative.dev/eventing-kafka-broker/control-plane/pkg/apis/internals/kafka/eventing"
+	internalscg "knative.dev/eventing-kafka-broker/control-plane/pkg/apis/internals/kafka/eventing/v1alpha1"
+	"knative.dev/eventing-kafka-broker/control-plane/pkg/config"
+	"knative.dev/eventing-kafka-broker/control-plane/pkg/kafka"
+	"knative.dev/eventing-kafka-broker/control-plane/pkg/reconciler/base"
+	brokerreconciler "knative.dev/eventing-kafka-broker/control-plane/pkg/reconciler/broker"
+
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	clientgotesting "k8s.io/client-go/testing"
 	eventing "knative.dev/eventing/pkg/apis/eventing/v1"
-	"knative.dev/eventing/pkg/apis/feature"
 	eventingclient "knative.dev/eventing/pkg/client/injection/client/fake"
 	triggerreconciler "knative.dev/eventing/pkg/client/injection/reconciler/eventing/v1/trigger"
 	reconcilertesting "knative.dev/eventing/pkg/reconciler/testing/v1"
+	"knative.dev/pkg/apis"
+	kubeclient "knative.dev/pkg/client/injection/kube/client"
+	"knative.dev/pkg/controller"
+	"knative.dev/pkg/kmeta"
+	"knative.dev/pkg/logging"
+	. "knative.dev/pkg/reconciler/testing"
 
-	internals "knative.dev/eventing-kafka-broker/control-plane/pkg/apis/internals/kafka/eventing"
-	"knative.dev/eventing-kafka-broker/control-plane/pkg/config"
-	"knative.dev/eventing-kafka-broker/control-plane/pkg/contract"
-	"knative.dev/eventing-kafka-broker/control-plane/pkg/kafka"
-	kafkatesting "knative.dev/eventing-kafka-broker/control-plane/pkg/kafka/testing"
-	brokerreconciler "knative.dev/eventing-kafka-broker/control-plane/pkg/reconciler/broker"
-
-	"knative.dev/eventing-kafka-broker/control-plane/pkg/receiver"
-	"knative.dev/eventing-kafka-broker/control-plane/pkg/reconciler/base"
 	. "knative.dev/eventing-kafka-broker/control-plane/pkg/reconciler/testing"
+
+	fakeconsumergroupinformer "knative.dev/eventing-kafka-broker/control-plane/pkg/client/internals/kafka/injection/client/fake"
+)
+
+const (
+	// name of the trigger under test
+	triggerName = "test-trigger"
+	// namespace of the trigger under test
+	triggerNamespace = "test-namespace"
+
+	bootstrapServers = "kafka-1:9092,kafka-2:9093"
 )
 
 var DefaultEnv = &config.Env{
 	DataPlaneConfigMapNamespace:  "knative-eventing",
-	ContractConfigMapName:        "kafka-broker-brokers-triggers",
-	GeneralConfigMapName:         "kafka-broker-config",
 	DataPlaneConfigConfigMapName: "config-kafka-broker-data-plane",
-	IngressName:                  "kafka-broker-ingress",
-	SystemNamespace:              "knative-eventing",
 	ContractConfigMapFormat:      base.Json,
-	DefaultBackoffDelayMs:        1000,
+	GeneralConfigMapName:         "kafka-broker-config",
+	SystemNamespace:              "knative-eventing",
 }
 
 var (
-	finalizerUpdatedEvent = Eventf(
-		corev1.EventTypeNormal,
-		"FinalizerUpdate",
-		fmt.Sprintf(`Updated %q finalizers`, TriggerName),
-	)
-
 	url = &apis.URL{
 		Scheme: "http",
 		Host:   "localhost",
@@ -84,52 +76,17 @@ var (
 	}
 
 	exponential = eventingduck.BackoffPolicyExponential
-
-	bootstrapServers = "kafka-1:9092,kafka-2:9093"
 )
 
-type EgressBuilder struct {
-	*contract.Egress
-}
+func TestReconcileKind(t *testing.T) {
 
-func (b EgressBuilder) build(useDialectedFilters bool) *contract.Egress {
-	if useDialectedFilters {
-		b.Filter = nil
-	} else {
-		b.DialectedFilter = nil
-	}
-	return b.Egress
-}
+	env := DefaultEnv
 
-func TestTriggerReconciler(t *testing.T) {
-	eventing.RegisterAlternateBrokerConditionSet(base.IngressConditionSet)
-
-	t.Parallel()
-
-	for _, f := range Formats {
-		triggerReconciliation(t, f, *DefaultEnv, false)
-	}
-}
-
-func TestTriggerWithNewFiltersReconciler(t *testing.T) {
-	eventing.RegisterAlternateBrokerConditionSet(base.IngressConditionSet)
-
-	t.Parallel()
-
-	for _, f := range Formats {
-		triggerReconciliation(t, f, *DefaultEnv, true)
-	}
-}
-
-func triggerReconciliation(t *testing.T, format string, env config.Env, useNewFilters bool) {
-
-	testKey := fmt.Sprintf("%s/%s", TriggerNamespace, TriggerName)
-
-	env.ContractConfigMapFormat = format
+	testKey := fmt.Sprintf("%s/%s", triggerNamespace, triggerName)
 
 	table := TableTest{
 		{
-			Name: "Reconciled normal - no existing Trigger",
+			Name: "Reconciled normal",
 			Objects: []runtime.Object{
 				NewBroker(
 					BrokerReady,
@@ -137,207 +94,92 @@ func triggerReconciliation(t *testing.T, format string, env config.Env, useNewFi
 					WithBootstrapServerStatusAnnotation(bootstrapServers),
 				),
 				newTrigger(),
-				NewService(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-						},
-					},
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
 				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
 					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
 				),
 			},
 			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{
-				ConfigMapUpdate(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat, &contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-									Reference:     TriggerReference(),
-								},
-							},
-						},
-					},
-					Generation: 1,
-				}),
-				BrokerDispatcherPodUpdate(env.SystemNamespace, map[string]string{
-					base.VolumeGenerationAnnotationKey: "1",
-				}),
+			WantCreates: []runtime.Object{
+				NewConsumerGroup(
+					WithConsumerGroupName(TriggerUUID),
+					WithConsumerGroupNamespace(triggerNamespace),
+					WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+					WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+					WithConsumerGroupLabels(ConsumerTriggerLabel),
+					ConsumerGroupConsumerSpec(NewConsumerSpec(
+						ConsumerTopics(BrokerTopics[0]),
+						ConsumerConfigs(
+							ConsumerBootstrapServersConfig(bootstrapServers),
+							ConsumerGroupIdConfig(TriggerUUID),
+						),
+						ConsumerDelivery(NewConsumerSpecDelivery(internals.Unordered, ConsumerInitialOffset(sources.OffsetLatest))),
+						ConsumerFilters(NewConsumerSpecFilters()),
+						ConsumerReply(ConsumerTopicReply()),
+					)),
+				),
 			},
 			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
 				{
 					Object: newTrigger(
 						reconcilertesting.WithInitTriggerConditions,
 						reconcilertesting.WithTriggerSubscribed(),
-						withSubscriberURI,
-						reconcilertesting.WithTriggerDependencyReady(),
 						reconcilertesting.WithTriggerBrokerReady(),
-						withTriggerSubscriberResolvedSucceeded(contract.DeliveryOrder_UNORDERED),
-						reconcilertesting.WithTriggerDeadLetterSinkNotConfigured(),
+						withTriggerSubscriberResolvedSucceeded(),
+						reconcilertesting.WithTriggerDependencyUnknown("failed to reconcile consumer group", "consumer group is not ready"),
+						withDeadLetterSinkURI(""),
 					),
 				},
 			},
 		},
 		{
-			Name: "Reconciled normal - with Broker DLS",
+			Name: "Reconciled normal - fallback to broker delivery",
 			Objects: []runtime.Object{
 				NewBroker(
 					BrokerReady,
-					WithDelivery(),
+					WithRetry(pointer.Int32Ptr(10), &exponential, pointer.StringPtr("PT2S")),
 					WithTopicStatusAnnotation(BrokerTopic()),
 					WithBootstrapServerStatusAnnotation(bootstrapServers),
+				),
+				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
+					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
 				),
 				newTrigger(),
-				NewService(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-						},
-					},
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
-				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
-					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetEarliest),
-				),
 			},
 			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{
-				ConfigMapUpdate(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat, &contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-									EgressConfig:  &contract.EgressConfig{DeadLetter: ServiceURL},
-									Reference:     TriggerReference(),
-								},
-							},
-						},
-					},
-					Generation: 1,
-				}),
-				BrokerDispatcherPodUpdate(env.SystemNamespace, map[string]string{
-					base.VolumeGenerationAnnotationKey: "1",
-				}),
+			WantCreates: []runtime.Object{
+				NewConsumerGroup(
+					WithConsumerGroupName(TriggerUUID),
+					WithConsumerGroupNamespace(triggerNamespace),
+					WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+					WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+					WithConsumerGroupLabels(ConsumerTriggerLabel),
+					ConsumerGroupConsumerSpec(NewConsumerSpec(
+						ConsumerTopics(BrokerTopics[0]),
+						ConsumerConfigs(
+							ConsumerGroupIdConfig(TriggerUUID),
+							ConsumerBootstrapServersConfig(bootstrapServers),
+						),
+						ConsumerDelivery(NewConsumerSpecDelivery(
+							internals.Unordered,
+							NewConsumerRetry(10),
+							NewConsumerBackoffDelay("PT2S"),
+							NewConsumerBackoffPolicy(eventingduck.BackoffPolicyExponential),
+							ConsumerInitialOffset(sources.OffsetLatest),
+						)),
+						ConsumerFilters(NewConsumerSpecFilters()),
+						ConsumerReply(ConsumerTopicReply()),
+					)),
+				),
 			},
 			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
 				{
 					Object: newTrigger(
 						reconcilertesting.WithInitTriggerConditions,
 						reconcilertesting.WithTriggerSubscribed(),
-						withSubscriberURI,
-						reconcilertesting.WithTriggerDependencyReady(),
 						reconcilertesting.WithTriggerBrokerReady(),
-						withTriggerSubscriberResolvedSucceeded(contract.DeliveryOrder_UNORDERED),
-						withDeadLetterSinkURI(ServiceURL),
-					),
-				},
-			},
-		},
-		{
-			Name: "Reconciled normal - with Trigger DLS",
-			Objects: []runtime.Object{
-				NewBroker(
-					BrokerReady,
-					WithTopicStatusAnnotation(BrokerTopic()),
-					WithBootstrapServerStatusAnnotation(bootstrapServers),
-				),
-				newTrigger(withDelivery),
-				NewService(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-						},
-					},
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
-				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
-					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
-				),
-			},
-			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{
-				ConfigMapUpdate(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat, &contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-									Reference:     TriggerReference(),
-									EgressConfig: &contract.EgressConfig{
-										DeadLetter:    url.String(),
-										Retry:         3,
-										BackoffPolicy: contract.BackoffPolicy_Exponential,
-										BackoffDelay:  uint64(time.Second.Milliseconds()),
-										Timeout:       uint64((time.Second * 2).Milliseconds()),
-									},
-								},
-							},
-						},
-					},
-					Generation: 1,
-				}),
-				BrokerDispatcherPodUpdate(env.SystemNamespace, map[string]string{
-					base.VolumeGenerationAnnotationKey: "1",
-				}),
-			},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						withDelivery,
-						reconcilertesting.WithInitTriggerConditions,
-						reconcilertesting.WithTriggerSubscribed(),
-						withSubscriberURI,
-						reconcilertesting.WithTriggerDependencyReady(),
-						reconcilertesting.WithTriggerBrokerReady(),
-						withTriggerSubscriberResolvedSucceeded(contract.DeliveryOrder_UNORDERED),
-						withDeadLetterSinkURI(url.String()),
+						withTriggerSubscriberResolvedSucceeded(),
+						reconcilertesting.WithTriggerDependencyUnknown("failed to reconcile consumer group", "consumer group is not ready"),
+						withDeadLetterSinkURI(""),
 					),
 				},
 			},
@@ -350,64 +192,41 @@ func triggerReconciliation(t *testing.T, format string, env config.Env, useNewFi
 					WithTopicStatusAnnotation(BrokerTopic()),
 					WithBootstrapServerStatusAnnotation(bootstrapServers),
 				),
-				newTrigger(reconcilertesting.WithAnnotation(deliveryOrderAnnotation, string(internals.Ordered))),
-				NewService(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-						},
-					},
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
 				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
 					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
 				),
+				newTrigger(reconcilertesting.WithAnnotation(deliveryOrderAnnotation, string(internals.Ordered))),
 			},
 			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{
-				ConfigMapUpdate(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat, &contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-									Reference:     TriggerReference(),
-									DeliveryOrder: contract.DeliveryOrder_ORDERED,
-								},
-							},
-						},
-					},
-					Generation: 1,
-				}),
-				BrokerDispatcherPodUpdate(env.SystemNamespace, map[string]string{
-					base.VolumeGenerationAnnotationKey: "1",
-				}),
+			WantCreates: []runtime.Object{
+				NewConsumerGroup(
+					WithConsumerGroupName(TriggerUUID),
+					WithConsumerGroupNamespace(triggerNamespace),
+					WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+					WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+					WithConsumerGroupLabels(ConsumerTriggerLabel),
+					ConsumerGroupConsumerSpec(NewConsumerSpec(
+						ConsumerTopics(BrokerTopics[0]),
+						ConsumerConfigs(
+							ConsumerGroupIdConfig(TriggerUUID),
+							ConsumerBootstrapServersConfig(bootstrapServers),
+						),
+						ConsumerDelivery(NewConsumerSpecDelivery(internals.Ordered, ConsumerInitialOffset(sources.OffsetLatest))),
+						ConsumerFilters(NewConsumerSpecFilters()),
+						ConsumerReply(ConsumerTopicReply()),
+					)),
+				),
 			},
 			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
 				{
 					Object: newTrigger(
 						reconcilertesting.WithInitTriggerConditions,
 						reconcilertesting.WithTriggerSubscribed(),
-						withSubscriberURI,
-						reconcilertesting.WithTriggerDependencyReady(),
 						reconcilertesting.WithTriggerBrokerReady(),
-						withTriggerSubscriberResolvedSucceeded(contract.DeliveryOrder_ORDERED),
+						withTriggerSubscriberResolvedSucceeded(),
+						reconcilertesting.WithTriggerDependencyUnknown("failed to reconcile consumer group", "consumer group is not ready"),
+						withDeadLetterSinkURI(""),
 						reconcilertesting.WithAnnotation(deliveryOrderAnnotation, string(internals.Ordered)),
-						reconcilertesting.WithTriggerDeadLetterSinkNotConfigured(),
 					),
 				},
 			},
@@ -420,348 +239,47 @@ func triggerReconciliation(t *testing.T, format string, env config.Env, useNewFi
 					WithTopicStatusAnnotation(BrokerTopic()),
 					WithBootstrapServerStatusAnnotation(bootstrapServers),
 				),
+				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
+					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
+				),
 				newTrigger(reconcilertesting.WithAnnotation(deliveryOrderAnnotation, string(internals.Unordered))),
-				NewService(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-						},
-					},
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
-				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
-					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
-				),
 			},
 			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{
-				ConfigMapUpdate(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat, &contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-									Reference:     TriggerReference(),
-									DeliveryOrder: contract.DeliveryOrder_UNORDERED,
-								},
-							},
-						},
-					},
-					Generation: 1,
-				}),
-				BrokerDispatcherPodUpdate(env.SystemNamespace, map[string]string{
-					base.VolumeGenerationAnnotationKey: "1",
-				}),
-			},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						reconcilertesting.WithInitTriggerConditions,
-						reconcilertesting.WithTriggerSubscribed(),
-						withSubscriberURI,
-						reconcilertesting.WithTriggerDependencyReady(),
-						reconcilertesting.WithTriggerBrokerReady(),
-						withTriggerSubscriberResolvedSucceeded(contract.DeliveryOrder_UNORDERED),
-						reconcilertesting.WithAnnotation(deliveryOrderAnnotation, string(internals.Unordered)),
-						reconcilertesting.WithTriggerDeadLetterSinkNotConfigured(),
-					),
-				},
-			},
-		},
-		{
-			Name: "Reconciled normal - Trigger delivery",
-			Objects: []runtime.Object{
-				NewBroker(
-					BrokerReady,
-					WithTopicStatusAnnotation(BrokerTopic()),
-					WithBootstrapServerStatusAnnotation(bootstrapServers),
-				),
-				newTrigger(withDelivery),
-				NewService(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-						},
-					},
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
-				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
-					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
-				),
-			},
-			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{
-				ConfigMapUpdate(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat, &contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-									Reference:     TriggerReference(),
-									EgressConfig: &contract.EgressConfig{
-										DeadLetter:    url.String(),
-										Retry:         3,
-										BackoffPolicy: contract.BackoffPolicy_Exponential,
-										BackoffDelay:  uint64(time.Second.Milliseconds()),
-										Timeout:       uint64((time.Second * 2).Milliseconds()),
-									},
-								},
-							},
-						},
-					},
-					Generation: 1,
-				}),
-				BrokerDispatcherPodUpdate(env.SystemNamespace, map[string]string{
-					base.VolumeGenerationAnnotationKey: "1",
-				}),
-			},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						withDelivery,
-						reconcilertesting.WithInitTriggerConditions,
-						reconcilertesting.WithTriggerSubscribed(),
-						withSubscriberURI,
-						reconcilertesting.WithTriggerDependencyReady(),
-						reconcilertesting.WithTriggerBrokerReady(),
-						withTriggerSubscriberResolvedSucceeded(contract.DeliveryOrder_UNORDERED),
-						reconcilertesting.WithTriggerDeadLetterSinkResolvedSucceeded(),
-						reconcilertesting.WithTriggerStatusDeadLetterSinkURI("http://localhost/path"),
-					),
-				},
-			},
-		},
-		{
-			Name: "Reconciled normal - with existing Triggers and Brokers",
-			Objects: []runtime.Object{
-				NewBroker(
-					BrokerReady,
-					WithTopicStatusAnnotation(BrokerTopic()),
-					WithBootstrapServerStatusAnnotation(bootstrapServers),
-				),
-				newTrigger(),
-				NewService(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID + "z",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-									Reference:     TriggerReference(),
-								},
-							},
-						},
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								EgressBuilder{
-									&contract.Egress{
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source_value",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source_value",
-													},
-												},
-											}},
-										},
-										Destination:   ServiceURL,
-										ConsumerGroup: TriggerUUID + "a",
-										Uid:           TriggerUUID + "a",
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source_value",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source_value",
-													},
-												},
-											}},
-										},
-										Destination:   ServiceURL,
-										ConsumerGroup: TriggerUUID,
-										Uid:           TriggerUUID,
-										Reference:     TriggerReference(),
-									},
-								}.build(useNewFilters),
-							},
-						},
-					},
-					Generation: 2,
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
-				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
-					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
-				),
-			},
-			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{
-				ConfigMapUpdate(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat, &contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID + "z",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-									Reference:     TriggerReference(),
-								},
-							},
-						},
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								EgressBuilder{
-									&contract.Egress{
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source_value",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source_value",
-													},
-												},
-											}},
-										},
-										Destination:   ServiceURL,
-										ConsumerGroup: TriggerUUID + "a",
-										Uid:           TriggerUUID + "a",
-									},
-								}.build(useNewFilters),
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-									Reference:     TriggerReference(),
-								},
-							},
-						},
-					},
-					Generation: 3,
-				}),
-				BrokerDispatcherPodUpdate(env.SystemNamespace, map[string]string{
-					base.VolumeGenerationAnnotationKey: "3",
-				}),
-			},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						reconcilertesting.WithInitTriggerConditions,
-						reconcilertesting.WithTriggerSubscribed(),
-						withSubscriberURI,
-						reconcilertesting.WithTriggerDependencyReady(),
-						reconcilertesting.WithTriggerBrokerReady(),
-						withTriggerSubscriberResolvedSucceeded(contract.DeliveryOrder_UNORDERED),
-						reconcilertesting.WithTriggerDeadLetterSinkNotConfigured(),
-					),
-				},
-			},
-		},
-		{
-			Name: "Config map not found - broker not found in config map",
-			Objects: []runtime.Object{
-				NewBroker(
-					BrokerReady,
-					WithTopicStatusAnnotation(BrokerTopic()),
-					WithBootstrapServerStatusAnnotation(bootstrapServers),
-				),
-				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
-					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
-				),
-				newTrigger(),
-				NewService(),
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			Key:                     testKey,
-			WantErr:                 true,
-			SkipNamespaceValidation: true, // WantCreates compare the broker namespace with configmap namespace, so skip it
 			WantCreates: []runtime.Object{
-				NewConfigMapWithBinaryData(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, nil),
-			},
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-				Eventf(
-					corev1.EventTypeWarning,
-					"InternalError",
-					fmt.Sprintf(
-						"broker not found in data plane config map %s",
-						env.DataPlaneConfigMapAsString(),
-					),
+				NewConsumerGroup(
+					WithConsumerGroupName(TriggerUUID),
+					WithConsumerGroupNamespace(triggerNamespace),
+					WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+					WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+					WithConsumerGroupLabels(ConsumerTriggerLabel),
+					ConsumerGroupConsumerSpec(NewConsumerSpec(
+						ConsumerTopics(BrokerTopics[0]),
+						ConsumerConfigs(
+							ConsumerGroupIdConfig(TriggerUUID),
+							ConsumerBootstrapServersConfig(bootstrapServers),
+						),
+						ConsumerDelivery(NewConsumerSpecDelivery(internals.Unordered, ConsumerInitialOffset(sources.OffsetLatest))),
+						ConsumerFilters(NewConsumerSpecFilters()),
+						ConsumerReply(ConsumerTopicReply()),
+					)),
 				),
 			},
 			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
 				{
 					Object: newTrigger(
 						reconcilertesting.WithInitTriggerConditions,
-						reconcilertesting.WithTriggerBrokerFailed(
-							"Broker not found in data plane map",
-							fmt.Sprintf("config map: %s", env.DataPlaneConfigMapAsString()),
-						),
+						reconcilertesting.WithTriggerSubscribed(),
+						reconcilertesting.WithTriggerBrokerReady(),
+						withTriggerSubscriberResolvedSucceeded(),
+						reconcilertesting.WithTriggerDependencyUnknown("failed to reconcile consumer group", "consumer group is not ready"),
+						withDeadLetterSinkURI(""),
+						reconcilertesting.WithAnnotation(deliveryOrderAnnotation, string(internals.Unordered)),
 					),
 				},
 			},
 		},
 		{
-			Name: "Empty data plane config map",
+			Name: "Reconciled normal - Trigger with invalid delivery",
 			Objects: []runtime.Object{
 				NewBroker(
 					BrokerReady,
@@ -771,23 +289,17 @@ func triggerReconciliation(t *testing.T, format string, env config.Env, useNewFi
 				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
 					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
 				),
-				newTrigger(),
-				NewService(),
-				NewConfigMapWithBinaryData(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, nil),
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
+				newTrigger(reconcilertesting.WithAnnotation(deliveryOrderAnnotation, "invalid")),
 			},
 			Key:     testKey,
 			WantErr: true,
 			WantEvents: []string{
-				finalizerUpdatedEvent,
 				Eventf(
 					corev1.EventTypeWarning,
 					"InternalError",
 					fmt.Sprintf(
-						"broker not found in data plane config map %s",
-						env.DataPlaneConfigMapAsString(),
+						"invalid annotation %s value: invalid. Allowed values [ \"ordered\" | \"unordered\" ]",
+						deliveryOrderAnnotation,
 					),
 				),
 			},
@@ -795,25 +307,473 @@ func triggerReconciliation(t *testing.T, format string, env config.Env, useNewFi
 				{
 					Object: newTrigger(
 						reconcilertesting.WithInitTriggerConditions,
-						reconcilertesting.WithTriggerBrokerFailed(
-							"Broker not found in data plane map",
-							fmt.Sprintf("config map: %s", env.DataPlaneConfigMapAsString()),
-						),
+						reconcilertesting.WithTriggerBrokerReady(),
+						reconcilertesting.WithTriggerDependencyFailed("failed to reconcile consumer group", "invalid annotation kafka.eventing.knative.dev/delivery.order value: invalid. Allowed values [ \"ordered\" | \"unordered\" ]"),
+						reconcilertesting.WithAnnotation(deliveryOrderAnnotation, "invalid"),
 					),
 				},
 			},
 		},
 		{
-			Name: "Broker not found, no broker in config map",
+			Name: "Reconciled normal - existing cg with update",
+			Objects: []runtime.Object{
+				NewBroker(
+					BrokerReady,
+					WithTopicStatusAnnotation(BrokerTopic()),
+					WithBootstrapServerStatusAnnotation(bootstrapServers),
+				),
+				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
+					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
+				),
+				newTrigger(),
+				NewConsumerGroup(
+					WithConsumerGroupName(TriggerUUID),
+					WithConsumerGroupNamespace(triggerNamespace),
+					WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+					WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+					WithConsumerGroupLabels(ConsumerTriggerLabel),
+					ConsumerGroupReady,
+				),
+			},
+			Key:         testKey,
+			WantCreates: []runtime.Object{},
+			WantUpdates: []clientgotesting.UpdateActionImpl{
+				{
+					Object: NewConsumerGroup(
+						WithConsumerGroupName(TriggerUUID),
+						WithConsumerGroupNamespace(triggerNamespace),
+						WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+						WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+						WithConsumerGroupLabels(ConsumerTriggerLabel),
+						ConsumerGroupConsumerSpec(NewConsumerSpec(
+							ConsumerTopics(BrokerTopics[0]),
+							ConsumerConfigs(
+								ConsumerGroupIdConfig(TriggerUUID),
+								ConsumerBootstrapServersConfig(bootstrapServers),
+							),
+							ConsumerDelivery(NewConsumerSpecDelivery(internals.Unordered, ConsumerInitialOffset(sources.OffsetLatest))),
+							ConsumerFilters(NewConsumerSpecFilters()),
+							ConsumerReply(ConsumerTopicReply()),
+						)),
+						ConsumerGroupReady,
+					),
+				},
+			},
+			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
+				{
+					Object: newTrigger(
+						reconcilertesting.WithInitTriggerConditions,
+						reconcilertesting.WithTriggerSubscribed(),
+						reconcilertesting.WithTriggerBrokerReady(),
+						withTriggerSubscriberResolvedSucceeded(),
+						reconcilertesting.WithTriggerDependencyReady(),
+						withDeadLetterSinkURI(""),
+					),
+				},
+			},
+		},
+		{
+			Name: "Reconciled normal - existing cg with autoscaling annotations",
+			Objects: []runtime.Object{
+				NewBroker(
+					BrokerReady,
+					WithTopicStatusAnnotation(BrokerTopic()),
+					WithBootstrapServerStatusAnnotation(bootstrapServers),
+				),
+				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
+					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
+				),
+				newTrigger(
+					WithAutoscalingAnnotationsTrigger(),
+				),
+				NewConsumerGroup(
+					WithConsumerGroupName(TriggerUUID),
+					WithConsumerGroupNamespace(triggerNamespace),
+					WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+					WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+					WithConsumerGroupLabels(ConsumerTriggerLabel),
+					ConsumerGroupReady,
+				),
+			},
+			Key:         testKey,
+			WantCreates: []runtime.Object{},
+			WantUpdates: []clientgotesting.UpdateActionImpl{
+				{
+					Object: NewConsumerGroup(
+						WithConsumerGroupName(TriggerUUID),
+						WithConsumerGroupNamespace(triggerNamespace),
+						WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+						WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+						WithConsumerGroupLabels(ConsumerTriggerLabel),
+						WithConsumerGroupAnnotations(ConsumerGroupAnnotations),
+						ConsumerGroupConsumerSpec(NewConsumerSpec(
+							ConsumerTopics(BrokerTopics[0]),
+							ConsumerConfigs(
+								ConsumerGroupIdConfig(TriggerUUID),
+								ConsumerBootstrapServersConfig(bootstrapServers),
+							),
+							ConsumerDelivery(NewConsumerSpecDelivery(internals.Unordered, ConsumerInitialOffset(sources.OffsetLatest))),
+							ConsumerFilters(NewConsumerSpecFilters()),
+							ConsumerReply(ConsumerTopicReply()),
+						)),
+						ConsumerGroupReady,
+					),
+				},
+			},
+			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
+				{
+					Object: newTrigger(
+						reconcilertesting.WithInitTriggerConditions,
+						reconcilertesting.WithTriggerSubscribed(),
+						reconcilertesting.WithTriggerBrokerReady(),
+						withTriggerSubscriberResolvedSucceeded(),
+						reconcilertesting.WithTriggerDependencyReady(),
+						withDeadLetterSinkURI(""),
+						WithAutoscalingAnnotationsTrigger(),
+					),
+				},
+			},
+		},
+		{
+			Name: "Reconciled normal - existing cg with update but not ready",
+			Objects: []runtime.Object{
+				NewBroker(
+					BrokerReady,
+					WithTopicStatusAnnotation(BrokerTopic()),
+					WithBootstrapServerStatusAnnotation(bootstrapServers),
+				),
+				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
+					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
+				),
+				newTrigger(),
+				NewConsumerGroup(
+					WithConsumerGroupName(TriggerUUID),
+					WithConsumerGroupNamespace(triggerNamespace),
+					WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+					WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+					WithConsumerGroupLabels(ConsumerTriggerLabel),
+				),
+			},
+			Key:         testKey,
+			WantCreates: []runtime.Object{},
+			WantUpdates: []clientgotesting.UpdateActionImpl{
+				{
+					Object: NewConsumerGroup(
+						WithConsumerGroupName(TriggerUUID),
+						WithConsumerGroupNamespace(triggerNamespace),
+						WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+						WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+						WithConsumerGroupLabels(ConsumerTriggerLabel),
+						ConsumerGroupConsumerSpec(NewConsumerSpec(
+							ConsumerTopics(BrokerTopics[0]),
+							ConsumerConfigs(
+								ConsumerGroupIdConfig(TriggerUUID),
+								ConsumerBootstrapServersConfig(bootstrapServers),
+							),
+							ConsumerDelivery(NewConsumerSpecDelivery(internals.Unordered, ConsumerInitialOffset(sources.OffsetLatest))),
+							ConsumerFilters(NewConsumerSpecFilters()),
+							ConsumerReply(ConsumerTopicReply()),
+						)),
+					),
+				},
+			},
+			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
+				{
+					Object: newTrigger(
+						reconcilertesting.WithInitTriggerConditions,
+						reconcilertesting.WithTriggerSubscribed(),
+						reconcilertesting.WithTriggerBrokerReady(),
+						withTriggerSubscriberResolvedSucceeded(),
+						reconcilertesting.WithTriggerDependencyUnknown("failed to reconcile consumer group", "consumer group is not ready"),
+						withDeadLetterSinkURI(""),
+					),
+				},
+			},
+		},
+		{
+			Name: "Reconciled normal - with auth - SASL",
+			Objects: []runtime.Object{
+				NewBroker(
+					BrokerReady,
+					WithTopicStatusAnnotation(BrokerTopic()),
+					WithBootstrapServerStatusAnnotation(bootstrapServers),
+					WithSecretStatusAnnotation("secret-1"),
+				),
+				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
+					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
+				),
+				newTrigger(),
+				NewConsumerGroup(
+					WithConsumerGroupName(TriggerUUID),
+					WithConsumerGroupNamespace(triggerNamespace),
+					WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+					WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+					WithConsumerGroupLabels(ConsumerTriggerLabel),
+					ConsumerGroupConsumerSpec(NewConsumerSpec(
+						ConsumerTopics(ChannelTopic()),
+						ConsumerConfigs(
+							ConsumerGroupIdConfig(TriggerUUID),
+							ConsumerBootstrapServersConfig(bootstrapServers),
+						),
+						ConsumerDelivery(NewConsumerSpecDelivery(internals.Ordered)),
+						ConsumerSubscriber(NewConsumerSpecSubscriber(Subscription1URI)),
+						ConsumerAuth(&internalscg.Auth{
+							SecretSpec: &internalscg.SecretSpec{
+								Ref: &internalscg.SecretReference{
+									Name:      "secret-1",
+									Namespace: ConfigMapNamespace,
+								},
+							},
+						}),
+					)),
+				),
+				NewLegacySASLSecret(ConfigMapNamespace, "secret-1"),
+			},
+			Key: testKey,
+			WantUpdates: []clientgotesting.UpdateActionImpl{
+				{
+					Object: NewConsumerGroup(
+						WithConsumerGroupName(TriggerUUID),
+						WithConsumerGroupNamespace(triggerNamespace),
+						WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+						WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+						WithConsumerGroupLabels(ConsumerTriggerLabel),
+						ConsumerGroupConsumerSpec(NewConsumerSpec(
+							ConsumerTopics(BrokerTopics[0]),
+							ConsumerConfigs(
+								ConsumerGroupIdConfig(TriggerUUID),
+								ConsumerBootstrapServersConfig(bootstrapServers),
+							),
+							ConsumerDelivery(NewConsumerSpecDelivery(internals.Unordered, ConsumerInitialOffset(sources.OffsetLatest))),
+							ConsumerFilters(NewConsumerSpecFilters()),
+							ConsumerReply(ConsumerTopicReply()),
+							ConsumerAuth(&internalscg.Auth{
+								SecretSpec: &internalscg.SecretSpec{
+									Ref: &internalscg.SecretReference{
+										Name:      "secret-1",
+										Namespace: ConfigMapNamespace,
+									},
+								},
+							}),
+						)),
+					),
+				},
+			},
+			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
+				{
+					Object: newTrigger(
+						reconcilertesting.WithInitTriggerConditions,
+						reconcilertesting.WithTriggerSubscribed(),
+						reconcilertesting.WithTriggerBrokerReady(),
+						withTriggerSubscriberResolvedSucceeded(),
+						reconcilertesting.WithTriggerDependencyUnknown("failed to reconcile consumer group", "consumer group is not ready"),
+						withDeadLetterSinkURI(""),
+					),
+				},
+			},
+		},
+		{
+			Name: "Reconciled normal - existing cg without update",
+			Objects: []runtime.Object{
+				NewBroker(
+					BrokerReady,
+					WithTopicStatusAnnotation(BrokerTopic()),
+					WithBootstrapServerStatusAnnotation(bootstrapServers),
+				),
+				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
+					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
+				),
+				newTrigger(),
+				NewConsumerGroup(
+					WithConsumerGroupName(TriggerUUID),
+					WithConsumerGroupNamespace(triggerNamespace),
+					WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+					WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+					WithConsumerGroupLabels(ConsumerTriggerLabel),
+					WithConsumerGroupAnnotations(ConsumerGroupAnnotations),
+					ConsumerGroupConsumerSpec(NewConsumerSpec(
+						ConsumerTopics(BrokerTopics[0]),
+						ConsumerConfigs(
+							ConsumerGroupIdConfig(TriggerUUID),
+							ConsumerBootstrapServersConfig(bootstrapServers),
+						),
+						ConsumerDelivery(NewConsumerSpecDelivery(internals.Unordered, ConsumerInitialOffset(sources.OffsetLatest))),
+						ConsumerFilters(NewConsumerSpecFilters()),
+						ConsumerReply(ConsumerTopicReply()),
+					)),
+					ConsumerGroupReady,
+					ConsumerGroupReplicas(1),
+				),
+			},
+			Key: testKey,
+			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
+				{
+					Object: newTrigger(
+						reconcilertesting.WithInitTriggerConditions,
+						reconcilertesting.WithTriggerSubscribed(),
+						reconcilertesting.WithTriggerBrokerReady(),
+						withTriggerSubscriberResolvedSucceeded(),
+						reconcilertesting.WithTriggerDependencyReady(),
+						withDeadLetterSinkURI(""),
+					),
+				},
+			},
+		},
+		{
+			Name: "Reconciled normal - existing cg without update but not ready",
+			Objects: []runtime.Object{
+				NewBroker(
+					BrokerReady,
+					WithTopicStatusAnnotation(BrokerTopic()),
+					WithBootstrapServerStatusAnnotation(bootstrapServers),
+				),
+				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
+					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
+				),
+				newTrigger(),
+				NewConsumerGroup(
+					WithConsumerGroupName(TriggerUUID),
+					WithConsumerGroupNamespace(triggerNamespace),
+					WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+					WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+					WithConsumerGroupLabels(ConsumerTriggerLabel),
+					WithConsumerGroupAnnotations(ConsumerGroupAnnotations),
+					ConsumerGroupConsumerSpec(NewConsumerSpec(
+						ConsumerTopics(BrokerTopics[0]),
+						ConsumerConfigs(
+							ConsumerGroupIdConfig(TriggerUUID),
+							ConsumerBootstrapServersConfig(bootstrapServers),
+						),
+						ConsumerDelivery(NewConsumerSpecDelivery(internals.Unordered, ConsumerInitialOffset(sources.OffsetLatest))),
+						ConsumerFilters(NewConsumerSpecFilters()),
+						ConsumerReply(ConsumerTopicReply()),
+					)),
+					ConsumerGroupReplicas(1),
+				),
+			},
+			Key: testKey,
+			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
+				{
+					Object: newTrigger(
+						reconcilertesting.WithInitTriggerConditions,
+						reconcilertesting.WithTriggerSubscribed(),
+						reconcilertesting.WithTriggerBrokerReady(),
+						withTriggerSubscriberResolvedSucceeded(),
+						reconcilertesting.WithTriggerDependencyUnknown("failed to reconcile consumer group", "consumer group is not ready"),
+						withDeadLetterSinkURI(""),
+					),
+				},
+			},
+		},
+		{
+			Name: "Reconciled normal - existing cg but failed",
+			Objects: []runtime.Object{
+				NewBroker(
+					BrokerReady,
+					WithTopicStatusAnnotation(BrokerTopic()),
+					WithBootstrapServerStatusAnnotation(bootstrapServers),
+				),
+				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
+					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
+				),
+				newTrigger(),
+				NewConsumerGroup(
+					WithConsumerGroupName(TriggerUUID),
+					WithConsumerGroupNamespace(triggerNamespace),
+					WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+					WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+					WithConsumerGroupLabels(ConsumerTriggerLabel),
+					WithConsumerGroupAnnotations(ConsumerGroupAnnotations),
+					ConsumerGroupConsumerSpec(NewConsumerSpec(
+						ConsumerTopics(BrokerTopics[0]),
+						ConsumerConfigs(
+							ConsumerGroupIdConfig(TriggerUUID),
+							ConsumerBootstrapServersConfig(bootstrapServers),
+						),
+						ConsumerDelivery(NewConsumerSpecDelivery(internals.Unordered, ConsumerInitialOffset(sources.OffsetLatest))),
+						ConsumerFilters(NewConsumerSpecFilters()),
+						ConsumerReply(ConsumerTopicReply()),
+					)),
+					WithConsumerGroupFailed("failed", "failed"),
+					ConsumerGroupReplicas(1),
+				),
+			},
+			Key: testKey,
+			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
+				{
+					Object: newTrigger(
+						reconcilertesting.WithInitTriggerConditions,
+						reconcilertesting.WithTriggerSubscribed(),
+						reconcilertesting.WithTriggerBrokerReady(),
+						withTriggerSubscriberResolvedSucceeded(),
+						reconcilertesting.WithTriggerDependencyFailed("failed", "failed"),
+						withDeadLetterSinkURI(""),
+					),
+				},
+			},
+		},
+		{
+			Name: "Reconciled normal with dead letter sink uri",
+			Objects: []runtime.Object{
+				NewBroker(
+					BrokerReady,
+					WithTopicStatusAnnotation(BrokerTopic()),
+					WithBootstrapServerStatusAnnotation(bootstrapServers),
+				),
+				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
+					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
+				),
+				newTrigger(),
+				NewConsumerGroup(
+					WithConsumerGroupName(TriggerUUID),
+					WithConsumerGroupNamespace(triggerNamespace),
+					WithConsumerGroupOwnerRef(kmeta.NewControllerRef(newTrigger())),
+					WithConsumerGroupMetaLabels(OwnerAsTriggerLabel),
+					WithConsumerGroupLabels(ConsumerTriggerLabel),
+					ConsumerGroupConsumerSpec(NewConsumerSpec(
+						ConsumerTopics(BrokerTopics[0]),
+						ConsumerConfigs(
+							ConsumerGroupIdConfig(TriggerUUID),
+							ConsumerBootstrapServersConfig(bootstrapServers),
+						),
+						ConsumerDelivery(NewConsumerSpecDelivery(internals.Unordered, ConsumerInitialOffset(sources.OffsetLatest))),
+						ConsumerFilters(NewConsumerSpecFilters()),
+						ConsumerReply(ConsumerTopicReply()),
+					)),
+					WithDeadLetterSinkURI(url.String()),
+					ConsumerGroupReplicas(1),
+				),
+			},
+			Key: testKey,
+			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
+				{
+					Object: newTrigger(
+						reconcilertesting.WithInitTriggerConditions,
+						reconcilertesting.WithTriggerSubscribed(),
+						reconcilertesting.WithTriggerBrokerReady(),
+						withTriggerSubscriberResolvedSucceeded(),
+						reconcilertesting.WithTriggerDependencyUnknown("failed to reconcile consumer group", "consumer group is not ready"),
+						withDeadLetterSinkURI(url.String()),
+					),
+				},
+			},
+		},
+		{
+			Name: "Broker not found",
 			Objects: []runtime.Object{
 				newTrigger(),
 			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			Key: testKey,
+			Key:     testKey,
+			WantErr: true,
 			WantEvents: []string{
-				finalizerUpdatedEvent,
+				Eventf(
+					corev1.EventTypeWarning,
+					"InternalError",
+					fmt.Sprintf(
+						"failed to get broker: brokers.eventing.knative.dev \"%s\" not found",
+						BrokerName,
+					),
+				),
 			},
 			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
 				{
@@ -832,1235 +792,29 @@ func triggerReconciliation(t *testing.T, format string, env config.Env, useNewFi
 					StatusBrokerConfigNotParsed("wrong"),
 				),
 			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
 			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
 			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
 				{
 					Object: newTrigger(
 						reconcilertesting.WithInitTriggerConditions,
 						reconcilertesting.WithTriggerBrokerFailed("wrong", ""),
+						withTriggerReadyFailed("wrong", ""),
 					),
 				},
 			},
 		},
 		{
-			Name: "Broker deleted, no broker in config map",
+			Name: "Broker deleted",
 			Objects: []runtime.Object{
 				newTrigger(),
 				NewDeletedBroker(),
-				NewConfigMapFromContract(&contract.Contract{
-					Generation: 8,
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
 			},
 			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
 			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
 				{
 					Object: newTrigger(
 						reconcilertesting.WithInitTriggerConditions,
-					),
-				},
-			},
-		},
-		{
-			Name: "Broker deleted, no trigger in config map",
-			Objects: []runtime.Object{
-				newTrigger(),
-				NewDeletedBroker(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:    BrokerUUID,
-							Topics: []string{BrokerTopic()},
-						},
-					},
-					Generation: 8,
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						reconcilertesting.WithInitTriggerConditions,
-					),
-				},
-			},
-		},
-		{
-			Name: "Broker deleted, trigger in config map",
-			Objects: []runtime.Object{
-				newTrigger(),
-				NewDeletedBroker(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:    BrokerUUID,
-							Topics: []string{BrokerTopic()},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID + "a",
-									Uid:           TriggerUUID + "a",
-								},
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-									Reference:     TriggerReference(),
-								},
-							},
-						},
-					},
-					Generation: 8,
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{
-				ConfigMapUpdate(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat, &contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:    BrokerUUID,
-							Topics: []string{BrokerTopic()},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID + "a",
-									Uid:           TriggerUUID + "a",
-								},
-							},
-						},
-					},
-					Generation: 9,
-				}),
-				BrokerDispatcherPodUpdate(env.SystemNamespace, map[string]string{
-					base.VolumeGenerationAnnotationKey: "9",
-				}),
-			},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						reconcilertesting.WithInitTriggerConditions,
-					),
-				},
-			},
-		},
-		{
-			Name: "Reconciled normal - no op",
-			Objects: []runtime.Object{
-				NewBroker(
-					BrokerReady,
-					WithTopicStatusAnnotation(BrokerTopic()),
-					WithBootstrapServerStatusAnnotation(bootstrapServers),
-				),
-				newTrigger(
-					withFilters(
-						map[string]string{"type": "type1"},
-						[]eventing.SubscriptionsAPIFilter{
-							{
-								Exact: map[string]string{
-									"type": "type1",
-								},
-							},
-						},
-						useNewFilters,
-					),
-				),
-				NewService(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								EgressBuilder{
-									&contract.Egress{
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"type": "type1",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"type": "type1",
-													},
-												},
-											}},
-										},
-										Destination:   ServiceURL,
-										ConsumerGroup: TriggerUUID,
-										Uid:           TriggerUUID,
-										Reference:     TriggerReference(),
-									},
-								}.build(useNewFilters),
-							},
-						},
-					},
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
-				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
-					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
-				),
-			},
-			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						withFilters(
-							map[string]string{"type": "type1"},
-							[]eventing.SubscriptionsAPIFilter{
-								{
-									Exact: map[string]string{
-										"type": "type1",
-									},
-								},
-							},
-							useNewFilters,
-						),
-						reconcilertesting.WithInitTriggerConditions,
-						reconcilertesting.WithTriggerSubscribed(),
-						withSubscriberURI,
-						reconcilertesting.WithTriggerDependencyReady(),
-						reconcilertesting.WithTriggerBrokerReady(),
-						withTriggerSubscriberResolvedSucceeded(contract.DeliveryOrder_UNORDERED),
-						reconcilertesting.WithTriggerDeadLetterSinkNotConfigured(),
-					),
-				},
-			},
-		},
-		{
-			Name: "Reconciled normal - many Triggers - start",
-			Objects: []runtime.Object{
-				NewBroker(
-					BrokerReady,
-					WithTopicStatusAnnotation(BrokerTopic()),
-					WithBootstrapServerStatusAnnotation(bootstrapServers),
-				),
-				newTrigger(
-					withFilters(
-						map[string]string{"type": "type1"},
-						[]eventing.SubscriptionsAPIFilter{
-							{
-								Exact: map[string]string{
-									"type": "type1",
-								},
-							},
-						},
-						useNewFilters,
-					),
-				),
-				NewService(),
-				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
-					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
-				),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-									Reference:     TriggerReference(),
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/2",
-										ConsumerGroup: "2",
-										Uid:           "2",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source2",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source2",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/3",
-										ConsumerGroup: "3",
-										Uid:           "3",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source3",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source3",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-							},
-						},
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/2",
-										ConsumerGroup: "2",
-										Uid:           "2",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source2",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source2",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/3",
-										ConsumerGroup: "3",
-										Uid:           "3",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source3",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source3",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-							},
-						},
-					},
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
-			},
-			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{
-				ConfigMapUpdate(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat, &contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   ServiceURL,
-										ConsumerGroup: TriggerUUID,
-										Uid:           TriggerUUID,
-										Reference:     TriggerReference(),
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"type": "type1",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"type": "type1",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/2",
-										ConsumerGroup: "2",
-										Uid:           "2",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source2",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source2",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/3",
-										ConsumerGroup: "3",
-										Uid:           "3",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source3",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source3",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-							},
-						},
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/2",
-										ConsumerGroup: "2",
-										Uid:           "2",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source2",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source2",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/3",
-										ConsumerGroup: "3",
-										Uid:           "3",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source3",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source3",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-							},
-						},
-					},
-					Generation: 1,
-				}),
-				BrokerDispatcherPodUpdate(env.SystemNamespace, map[string]string{
-					base.VolumeGenerationAnnotationKey: "1",
-				}),
-			},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						withFilters(
-							map[string]string{"type": "type1"},
-							[]eventing.SubscriptionsAPIFilter{
-								{
-									Exact: map[string]string{
-										"type": "type1",
-									},
-								},
-							},
-							useNewFilters,
-						),
-						reconcilertesting.WithInitTriggerConditions,
-						reconcilertesting.WithTriggerSubscribed(),
-						withSubscriberURI,
-						reconcilertesting.WithTriggerDependencyReady(),
-						reconcilertesting.WithTriggerBrokerReady(),
-						withTriggerSubscriberResolvedSucceeded(contract.DeliveryOrder_UNORDERED),
-						reconcilertesting.WithTriggerDeadLetterSinkNotConfigured(),
-					),
-				},
-			},
-		},
-		{
-			Name: "Reconciled normal - many Triggers - end",
-			Objects: []runtime.Object{
-				NewBroker(
-					BrokerReady,
-					WithTopicStatusAnnotation(BrokerTopic()),
-					WithBootstrapServerStatusAnnotation(bootstrapServers),
-				),
-				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
-					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
-				),
-				newTrigger(
-					withFilters(
-						map[string]string{"ext": "extval"},
-						[]eventing.SubscriptionsAPIFilter{
-							{
-								Exact: map[string]string{
-									"ext": "extval",
-								},
-							},
-						},
-						useNewFilters,
-					),
-				),
-				NewService(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/2",
-										ConsumerGroup: "2",
-										Uid:           "2",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source2",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source2",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/3",
-										ConsumerGroup: "3",
-										Uid:           "3",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source3",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source3",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-							},
-						},
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/2",
-										ConsumerGroup: "2",
-										Uid:           "2",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source2",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source2",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/3",
-										ConsumerGroup: "3",
-										Uid:           "3",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source3",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source3",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-							},
-						},
-					},
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
-			},
-			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{
-				ConfigMapUpdate(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat, &contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/2",
-										ConsumerGroup: "2",
-										Uid:           "2",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source2",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source2",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/3",
-										ConsumerGroup: "3",
-										Uid:           "3",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source3",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source3",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-							},
-						},
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/2",
-										ConsumerGroup: "2",
-										Uid:           "2",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source2",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source2",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/3",
-										ConsumerGroup: "3",
-										Uid:           "3",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source3",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source3",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   ServiceURL,
-										ConsumerGroup: TriggerUUID,
-										Uid:           TriggerUUID,
-										Reference:     TriggerReference(),
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"ext": "extval",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"ext": "extval",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-							},
-						},
-					},
-					Generation: 1,
-				}),
-				BrokerDispatcherPodUpdate(env.SystemNamespace, map[string]string{
-					base.VolumeGenerationAnnotationKey: "1",
-				}),
-			},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						withFilters(
-							map[string]string{"ext": "extval"},
-							[]eventing.SubscriptionsAPIFilter{
-								{
-									Exact: map[string]string{
-										"ext": "extval",
-									},
-								},
-							},
-							useNewFilters,
-						),
-						reconcilertesting.WithInitTriggerConditions,
-						reconcilertesting.WithTriggerSubscribed(),
-						withSubscriberURI,
-						reconcilertesting.WithTriggerDependencyReady(),
-						reconcilertesting.WithTriggerBrokerReady(),
-						withTriggerSubscriberResolvedSucceeded(contract.DeliveryOrder_UNORDERED),
-						reconcilertesting.WithTriggerDeadLetterSinkNotConfigured(),
-					),
-				},
-			},
-		},
-		{
-			Name: "Reconciled normal - many Triggers - middle",
-			Objects: []runtime.Object{
-				NewBroker(
-					BrokerReady,
-					WithTopicStatusAnnotation(BrokerTopic()),
-					WithBootstrapServerStatusAnnotation(bootstrapServers),
-				),
-				DataPlaneConfigMap(env.DataPlaneConfigMapNamespace, env.DataPlaneConfigConfigMapName, brokerreconciler.ConsumerConfigKey,
-					DataPlaneConfigInitialOffset(brokerreconciler.ConsumerConfigKey, sources.OffsetLatest),
-				),
-				newTrigger(),
-				NewService(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/2",
-										ConsumerGroup: "2",
-										Uid:           "2",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source2",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source2",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/3",
-										ConsumerGroup: "3",
-										Uid:           "3",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source3",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source3",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-							},
-						},
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/2",
-										ConsumerGroup: "2",
-										Uid:           "2",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source2",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source2",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   ServiceURL,
-										ConsumerGroup: TriggerUUID,
-										Uid:           TriggerUUID,
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source2",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source2",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/3",
-										ConsumerGroup: "3",
-										Uid:           "3",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source3",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source3",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-							},
-						},
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/2",
-										ConsumerGroup: "2",
-										Uid:           "2",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source2",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source2",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/3",
-										ConsumerGroup: "3",
-										Uid:           "3",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source3",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source3",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-							},
-						},
-					},
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
-			},
-			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{
-				ConfigMapUpdate(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat, &contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/2",
-										ConsumerGroup: "2",
-										Uid:           "2",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source2",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source2",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/3",
-										ConsumerGroup: "3",
-										Uid:           "3",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source3",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source3",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-							},
-						},
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/2",
-										ConsumerGroup: "2",
-										Uid:           "2",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source2",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source2",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-									Reference:     TriggerReference(),
-								},
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/3",
-										ConsumerGroup: "3",
-										Uid:           "3",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source3",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source3",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-							},
-						},
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/2",
-										ConsumerGroup: "2",
-										Uid:           "2",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source2",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source2",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-								EgressBuilder{
-									&contract.Egress{
-										Destination:   "http://example.com/3",
-										ConsumerGroup: "3",
-										Uid:           "3",
-										Filter: &contract.Filter{Attributes: map[string]string{
-											"source": "source3",
-										}},
-										DialectedFilter: []*contract.DialectedFilter{
-											{Filter: &contract.DialectedFilter_Exact{
-												Exact: &contract.Exact{
-													Attributes: map[string]string{
-														"source": "source3",
-													},
-												},
-											}},
-										},
-									},
-								}.build(useNewFilters),
-							},
-						},
-					},
-					Generation: 1,
-				}),
-				BrokerDispatcherPodUpdate(env.SystemNamespace, map[string]string{
-					base.VolumeGenerationAnnotationKey: "1",
-				}),
-			},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						reconcilertesting.WithInitTriggerConditions,
-						reconcilertesting.WithTriggerSubscribed(),
-						withSubscriberURI,
-						reconcilertesting.WithTriggerDependencyReady(),
-						reconcilertesting.WithTriggerBrokerReady(),
-						withTriggerSubscriberResolvedSucceeded(contract.DeliveryOrder_UNORDERED),
-						reconcilertesting.WithTriggerDeadLetterSinkNotConfigured(),
+						reconcilertesting.WithTriggerBrokerNotConfigured(),
 					),
 				},
 			},
@@ -2083,797 +837,23 @@ func triggerReconciliation(t *testing.T, format string, env config.Env, useNewFi
 					),
 				},
 			},
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
 		},
 	}
-
-	for i := range table {
-		table[i].Name = table[i].Name + " - " + format
-	}
-
-	if useNewFilters {
-		useTableWithFlags(t, table, &env, map[string]feature.Flag{feature.NewTriggerFilters: feature.Enabled})
-	} else {
-		useTable(t, table, &env)
-	}
-}
-
-func withDelivery(trigger *eventing.Trigger) {
-	trigger.Spec.Delivery = &eventingduck.DeliverySpec{
-		DeadLetterSink: &duckv1.Destination{URI: url},
-		Retry:          pointer.Int32(3),
-		BackoffPolicy:  &exponential,
-		BackoffDelay:   pointer.String("PT1S"),
-		Timeout:        pointer.String("PT2S"),
-	}
-}
-
-func TestTriggerFinalizer(t *testing.T) {
-
-	t.Parallel()
-
-	for _, f := range Formats {
-		triggerFinalizer(t, f, *DefaultEnv)
-	}
-
-}
-
-func triggerFinalizer(t *testing.T, format string, env config.Env) {
-
-	testKey := fmt.Sprintf("%s/%s", TriggerNamespace, TriggerName)
-
-	env.ContractConfigMapFormat = format
-
-	table := TableTest{
-		{
-			Name: "Broker deleted, trigger in config map",
-			Objects: []runtime.Object{
-				newTrigger(),
-				NewDeletedBroker(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:    BrokerUUID,
-							Topics: []string{BrokerTopic()},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID + "a",
-									Uid:           TriggerUUID + "a",
-								},
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-							},
-						},
-					},
-					Generation: 8,
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{
-				ConfigMapUpdate(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat, &contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:    BrokerUUID,
-							Topics: []string{BrokerTopic()},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID + "a",
-									Uid:           TriggerUUID + "a",
-								},
-							},
-						},
-					},
-					Generation: 9,
-				}),
-				BrokerDispatcherPodUpdate(env.SystemNamespace, map[string]string{
-					base.VolumeGenerationAnnotationKey: "9",
-				}),
-			},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						reconcilertesting.WithInitTriggerConditions,
-					),
-				},
-			},
-		},
-		{
-			Name: "Broker deleted, no trigger in config map",
-			Objects: []runtime.Object{
-				newTrigger(),
-				NewDeletedBroker(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:    BrokerUUID,
-							Topics: []string{BrokerTopic()},
-						},
-					},
-					Generation: 8,
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						reconcilertesting.WithInitTriggerConditions,
-					),
-				},
-			},
-		},
-		{
-			Name: "Broker deleted, no broker in config map",
-			Objects: []runtime.Object{
-				newTrigger(),
-				NewDeletedBroker(),
-				NewConfigMapFromContract(&contract.Contract{
-					Generation: 8,
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						reconcilertesting.WithInitTriggerConditions,
-					),
-				},
-			},
-		},
-		{
-			Name: "Reconciled normal - many Triggers - start",
-			Objects: []runtime.Object{
-				NewDeletedBroker(
-					BrokerReady,
-				),
-				newTrigger(),
-				NewService(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								{
-									Destination:   "http://example.com/2",
-									ConsumerGroup: "2",
-									Uid:           "2",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source2",
-									}},
-								},
-								{
-									Destination:   "http://example.com/3",
-									ConsumerGroup: "3",
-									Uid:           "3",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source3",
-									}},
-								},
-							},
-						},
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								{
-									Destination:   "http://example.com/2",
-									ConsumerGroup: "2",
-									Uid:           "2",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source2",
-									}},
-								},
-								{
-									Destination:   "http://example.com/3",
-									ConsumerGroup: "3",
-									Uid:           "3",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source3",
-									}},
-								},
-							},
-						},
-					},
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
-			},
-			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{
-				ConfigMapUpdate(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat, &contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   "http://example.com/3",
-									ConsumerGroup: "3",
-									Uid:           "3",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source3",
-									}},
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								{
-									Destination:   "http://example.com/2",
-									ConsumerGroup: "2",
-									Uid:           "2",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source2",
-									}},
-								},
-							},
-						},
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								{
-									Destination:   "http://example.com/2",
-									ConsumerGroup: "2",
-									Uid:           "2",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source2",
-									}},
-								},
-								{
-									Destination:   "http://example.com/3",
-									ConsumerGroup: "3",
-									Uid:           "3",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source3",
-									}},
-								},
-							},
-						},
-					},
-					Generation: 1,
-				}),
-				BrokerDispatcherPodUpdate(env.SystemNamespace, map[string]string{
-					base.VolumeGenerationAnnotationKey: "1",
-				}),
-			},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						reconcilertesting.WithInitTriggerConditions,
-					),
-				},
-			},
-		},
-		{
-			Name: "Reconciled normal - many Triggers - end",
-			Objects: []runtime.Object{
-				NewDeletedBroker(
-					BrokerReady,
-				),
-				newTrigger(),
-				NewService(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								{
-									Destination:   "http://example.com/2",
-									ConsumerGroup: "2",
-									Uid:           "2",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source2",
-									}},
-								},
-								{
-									Destination:   "http://example.com/3",
-									ConsumerGroup: "3",
-									Uid:           "3",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source3",
-									}},
-								},
-							},
-						},
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								{
-									Destination:   "http://example.com/2",
-									ConsumerGroup: "2",
-									Uid:           "2",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source2",
-									}},
-								},
-								{
-									Destination:   "http://example.com/3",
-									ConsumerGroup: "3",
-									Uid:           "3",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source3",
-									}},
-								},
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-							},
-						},
-					},
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
-			},
-			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{
-				ConfigMapUpdate(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat, &contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								{
-									Destination:   "http://example.com/2",
-									ConsumerGroup: "2",
-									Uid:           "2",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source2",
-									}},
-								},
-								{
-									Destination:   "http://example.com/3",
-									ConsumerGroup: "3",
-									Uid:           "3",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source3",
-									}},
-								},
-							},
-						},
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								{
-									Destination:   "http://example.com/2",
-									ConsumerGroup: "2",
-									Uid:           "2",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source2",
-									}},
-								},
-								{
-									Destination:   "http://example.com/3",
-									ConsumerGroup: "3",
-									Uid:           "3",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source3",
-									}},
-								},
-							},
-						},
-					},
-					Generation: 1,
-				}),
-				BrokerDispatcherPodUpdate(env.SystemNamespace, map[string]string{
-					base.VolumeGenerationAnnotationKey: "1",
-				}),
-			},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						reconcilertesting.WithInitTriggerConditions,
-					),
-				},
-			},
-		},
-		{
-			Name: "Reconciled normal - many Triggers - middle",
-			Objects: []runtime.Object{
-				NewDeletedBroker(
-					BrokerReady,
-				),
-				newTrigger(),
-				NewService(),
-				NewConfigMapFromContract(&contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								{
-									Destination:   "http://example.com/2",
-									ConsumerGroup: "2",
-									Uid:           "2",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source2",
-									}},
-								},
-								{
-									Destination:   "http://example.com/3",
-									ConsumerGroup: "3",
-									Uid:           "3",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source3",
-									}},
-								},
-							},
-						},
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								{
-									Destination:   "http://example.com/2",
-									ConsumerGroup: "2",
-									Uid:           "2",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source2",
-									}},
-								},
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/3",
-									ConsumerGroup: "3",
-									Uid:           "3",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source3",
-									}},
-								},
-							},
-						},
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								{
-									Destination:   "http://example.com/2",
-									ConsumerGroup: "2",
-									Uid:           "2",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source2",
-									}},
-								},
-								{
-									Destination:   "http://example.com/3",
-									ConsumerGroup: "3",
-									Uid:           "3",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source3",
-									}},
-								},
-							},
-						},
-					},
-				}, env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat),
-				BrokerDispatcherPod(env.SystemNamespace, nil),
-			},
-			Key: testKey,
-			WantEvents: []string{
-				finalizerUpdatedEvent,
-			},
-			WantPatches: []clientgotesting.PatchActionImpl{
-				patchFinalizers(),
-			},
-			WantUpdates: []clientgotesting.UpdateActionImpl{
-				ConfigMapUpdate(env.DataPlaneConfigMapNamespace, env.ContractConfigMapName, env.ContractConfigMapFormat, &contract.Contract{
-					Resources: []*contract.Resource{
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								{
-									Destination:   "http://example.com/2",
-									ConsumerGroup: "2",
-									Uid:           "2",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source2",
-									}},
-								},
-								{
-									Destination:   "http://example.com/3",
-									ConsumerGroup: "3",
-									Uid:           "3",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source3",
-									}},
-								},
-							},
-						},
-						{
-							Uid:     BrokerUUID,
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								{
-									Destination:   "http://example.com/2",
-									ConsumerGroup: "2",
-									Uid:           "2",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source2",
-									}},
-								},
-								{
-									Destination:   "http://example.com/3",
-									ConsumerGroup: "3",
-									Uid:           "3",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source3",
-									}},
-								},
-							},
-						},
-						{
-							Uid:     BrokerUUID + "a",
-							Topics:  []string{BrokerTopic()},
-							Ingress: &contract.Ingress{Path: receiver.Path(BrokerNamespace, BrokerName)},
-							Egresses: []*contract.Egress{
-								{
-									Destination:   ServiceURL,
-									ConsumerGroup: TriggerUUID,
-									Uid:           TriggerUUID,
-								},
-								{
-									Destination:   "http://example.com/1",
-									ConsumerGroup: "1",
-									Uid:           "1",
-								},
-								{
-									Destination:   "http://example.com/2",
-									ConsumerGroup: "2",
-									Uid:           "2",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source2",
-									}},
-								},
-								{
-									Destination:   "http://example.com/3",
-									ConsumerGroup: "3",
-									Uid:           "3",
-									Filter: &contract.Filter{Attributes: map[string]string{
-										"source": "source3",
-									}},
-								},
-							},
-						},
-					},
-					Generation: 1,
-				}),
-				BrokerDispatcherPodUpdate(env.SystemNamespace, map[string]string{
-					base.VolumeGenerationAnnotationKey: "1",
-				}),
-			},
-			WantStatusUpdates: []clientgotesting.UpdateActionImpl{
-				{
-					Object: newTrigger(
-						reconcilertesting.WithInitTriggerConditions,
-					),
-				},
-			},
-		},
-	}
-
-	useTable(t, table, &env)
-}
-
-func useTable(t *testing.T, table TableTest, env *config.Env) {
-	useTableWithFlags(t, table, env, nil)
-}
-
-func useTableWithFlags(t *testing.T, table TableTest, env *config.Env, flags feature.Flags) {
 
 	table.Test(t, NewFactory(env, func(ctx context.Context, listers *Listers, env *config.Env, row *TableRow) controller.Reconciler {
-
 		logger := logging.FromContext(ctx)
 
 		reconciler := &Reconciler{
-			Reconciler: &base.Reconciler{
-				KubeClient:                   kubeclient.Get(ctx),
-				PodLister:                    listers.GetPodLister(),
-				SecretLister:                 listers.GetSecretLister(),
-				DataPlaneConfigMapNamespace:  env.DataPlaneConfigMapNamespace,
-				ContractConfigMapName:        env.ContractConfigMapName,
-				ContractConfigMapFormat:      env.ContractConfigMapFormat,
-				DataPlaneConfigConfigMapName: env.DataPlaneConfigConfigMapName,
-				DataPlaneNamespace:           env.SystemNamespace,
-				DispatcherLabel:              base.BrokerDispatcherLabel,
-				ReceiverLabel:                base.BrokerReceiverLabel,
-			},
-			FlagsHolder: &FlagsHolder{
-				Flags: flags,
-			},
-			BrokerLister:              listers.GetBrokerLister(),
-			ConfigMapLister:           listers.GetConfigMapLister(),
-			EventingClient:            eventingclient.Get(ctx),
-			Resolver:                  nil,
-			Env:                       env,
-			BrokerClass:               kafka.BrokerClass,
-			DataPlaneConfigMapLabeler: base.NoopConfigmapOption,
-			InitOffsetsFunc: func(ctx context.Context, kafkaClient sarama.Client, kafkaAdminClient sarama.ClusterAdmin, topics []string, consumerGroup string) (int32, error) {
-				return 1, nil
-			},
-			NewKafkaClient: func(addrs []string, config *sarama.Config) (sarama.Client, error) {
-				return &kafkatesting.MockKafkaClient{}, nil
-			},
-			NewKafkaClusterAdminClient: func(_ []string, _ *sarama.Config) (sarama.ClusterAdmin, error) {
-				return &kafkatesting.MockKafkaClusterAdmin{
-					ExpectedTopicName: BrokerTopic(),
-					ExpectedTopics:    []string{BrokerTopic()},
-					ExpectedTopicsMetadataOnDescribeTopics: []*sarama.TopicMetadata{
-						{
-							Err:        0,
-							Name:       BrokerTopic(),
-							IsInternal: false,
-							Partitions: []*sarama.PartitionMetadata{{}},
-						},
-					},
-					T: t,
-				}, nil
-			},
+			BrokerLister:        listers.GetBrokerLister(),
+			ConfigMapLister:     listers.GetConfigMapLister(),
+			EventingClient:      eventingclient.Get(ctx),
+			Env:                 env,
+			BrokerClass:         kafka.BrokerClass,
+			ConsumerGroupLister: listers.GetConsumerGroupLister(),
+			InternalsClient:     fakeconsumergroupinformer.Get(ctx),
+			SecretLister:        listers.GetSecretLister(),
+			KubeClient:          kubeclient.Get(ctx),
 		}
-
-		reconciler.Resolver = resolver.NewURIResolverFromTracker(ctx, tracker.New(func(name types.NamespacedName) {}, 0))
 
 		return triggerreconciler.NewReconciler(
 			ctx,
@@ -2882,48 +862,23 @@ func useTableWithFlags(t *testing.T, table TableTest, env *config.Env, flags fea
 			listers.GetTriggerLister(),
 			controller.GetEventRecorder(ctx),
 			reconciler,
-			controller.Options{
-				FinalizerName:     FinalizerName,
-				AgentName:         ControllerAgentName,
-				SkipStatusUpdates: false,
-			},
 		)
 	}))
 }
 
-func newTrigger(options ...reconcilertesting.TriggerOption) runtime.Object {
+func newTrigger(options ...reconcilertesting.TriggerOption) *eventing.Trigger {
 	return reconcilertesting.NewTrigger(
-		TriggerName,
-		TriggerNamespace,
+		triggerName,
+		triggerNamespace,
 		BrokerName,
 		append(
 			options,
-			reconcilertesting.WithTriggerSubscriberURI(ServiceURL),
+			//reconcilertesting.WithTriggerSubscriberURI(ServiceURL),
 			func(t *eventing.Trigger) {
 				t.UID = TriggerUUID
 			},
 		)...,
 	)
-}
-
-func withFilters(attributes eventing.TriggerFilterAttributes, newFilters []eventing.SubscriptionsAPIFilter, withNewFilters bool) func(*eventing.Trigger) {
-	return func(e *eventing.Trigger) {
-		if withNewFilters {
-			e.Spec.Filters = newFilters
-		} else {
-			e.Spec.Filter = &eventing.TriggerFilter{
-				Attributes: attributes,
-			}
-		}
-	}
-}
-
-func withSubscriberURI(trigger *eventing.Trigger) {
-	u, err := apis.ParseURL(ServiceURL)
-	if err != nil {
-		panic(err)
-	}
-	trigger.Status.SubscriberURI = u
 }
 
 func withDeadLetterSinkURI(uri string) func(trigger *eventing.Trigger) {
@@ -2937,141 +892,18 @@ func withDeadLetterSinkURI(uri string) func(trigger *eventing.Trigger) {
 	}
 }
 
-func withTriggerSubscriberResolvedSucceeded(deliveryOrder contract.DeliveryOrder) func(*eventing.Trigger) {
+func withTriggerSubscriberResolvedSucceeded() func(*eventing.Trigger) {
 	return func(t *eventing.Trigger) {
-		t.GetConditionSet().Manage(&t.Status).MarkTrueWithReason(
+		t.GetConditionSet().Manage(&t.Status).MarkTrue(
 			eventing.TriggerConditionSubscriberResolved,
-			string(eventing.TriggerConditionSubscriberResolved),
-			fmt.Sprintf("Subscriber will receive events with the delivery order: %s", deliveryOrder.String()),
 		)
 	}
 }
 
-func patchFinalizers() clientgotesting.PatchActionImpl {
-	action := clientgotesting.PatchActionImpl{}
-	action.Name = TriggerName
-	action.Namespace = TriggerNamespace
-	patch := `{"metadata":{"finalizers":["` + FinalizerName + `"],"resourceVersion":""}}`
-	action.Patch = []byte(patch)
-	return action
-}
-
-func Test_deleteTrigger(t *testing.T) {
-	type args struct {
-		triggers []*contract.Egress
-		index    int
-	}
-	tests := []struct {
-		name string
-		args args
-		want []*contract.Egress
-	}{
-		{
-			name: "first element - alone",
-			args: args{
-				triggers: []*contract.Egress{
-					{
-						ConsumerGroup: "123",
-					},
-				},
-				index: 0,
-			},
-			want: nil,
-		},
-		{
-			name: "first element - with others",
-			args: args{
-				triggers: []*contract.Egress{
-					{
-						ConsumerGroup: "123",
-					},
-					{
-						ConsumerGroup: "1234",
-					},
-				},
-				index: 0,
-			},
-			want: []*contract.Egress{
-				{
-					ConsumerGroup: "1234",
-				},
-			},
-		},
-		{
-			name: "last element - with others",
-			args: args{
-				triggers: []*contract.Egress{
-					{
-						ConsumerGroup: "1",
-					},
-					{
-						ConsumerGroup: "2",
-					},
-					{
-						ConsumerGroup: "3",
-					},
-					{
-						ConsumerGroup: "4",
-					},
-				},
-				index: 3,
-			},
-			want: []*contract.Egress{
-				{
-					ConsumerGroup: "1",
-				},
-				{
-					ConsumerGroup: "2",
-				},
-				{
-					ConsumerGroup: "3",
-				},
-			},
-		},
-		{
-			name: "middle element - with others",
-			args: args{
-				triggers: []*contract.Egress{
-					{
-						ConsumerGroup: "1",
-					},
-					{
-						ConsumerGroup: "2",
-					},
-					{
-						ConsumerGroup: "3",
-					},
-					{
-						ConsumerGroup: "4",
-					},
-					{
-						ConsumerGroup: "5",
-					},
-				},
-				index: 2,
-			},
-			want: []*contract.Egress{
-				{
-					ConsumerGroup: "1",
-				},
-				{
-					ConsumerGroup: "2",
-				},
-				{
-					ConsumerGroup: "5",
-				},
-				{
-					ConsumerGroup: "4",
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := deleteTrigger(tt.args.triggers, tt.args.index)
-			if diff := cmp.Diff(tt.want, got, protocmp.Transform()); diff != "" {
-				t.Errorf("deleteTrigger() = %v, want %v (-want +got) %s", got, tt.want, diff)
-			}
-		})
+func withTriggerReadyFailed(reason, message string) func(*eventing.Trigger) {
+	return func(t *eventing.Trigger) {
+		t.GetConditionSet().Manage(&t.Status).MarkFalse(
+			eventing.TriggerConditionReady, reason, message,
+		)
 	}
 }
