+++ data-plane/config/broker/500-dispatcher.yaml
@@ -15,7 +15,7 @@
 # limitations under the License.
 
 apiVersion: apps/v1
-kind: Deployment
+kind: StatefulSet
 metadata:
   name: kafka-broker-dispatcher
   namespace: knative-eventing
@@ -23,6 +23,8 @@ metadata:
     app: kafka-broker-dispatcher
     kafka.eventing.knative.dev/release: devel
 spec:
+  serviceName: kafka-broker-dispatcher
+  podManagementPolicy: "Parallel"
   selector:
     matchLabels:
       app: kafka-broker-dispatcher
@@ -31,6 +33,7 @@ spec:
       name: kafka-broker-dispatcher
       labels:
         app: kafka-broker-dispatcher
+        app.kubernetes.io/component: kafka-dispatcher
         kafka.eventing.knative.dev/release: devel
     spec:
       # To avoid node becoming SPOF, spread our replicas to different nodes and zones.
@@ -61,13 +64,13 @@ spec:
             - mountPath: /etc/config
               name: config-kafka-broker-data-plane
               readOnly: true
-            - mountPath: /etc/brokers-triggers
-              name: kafka-broker-brokers-triggers
+            - mountPath: /etc/contract-resources
+              name: contract-resources
               readOnly: true
             - mountPath: /tmp
               name: cache
             - mountPath: /etc/logging
-              name: kafka-broker-config-logging
+              name: kafka-config-logging
               readOnly: true
             - mountPath: /etc/tracing
               name: config-tracing
@@ -90,7 +93,7 @@ spec:
             - name: WEBCLIENT_CONFIG_FILE_PATH
               value: /etc/config/config-kafka-broker-webclient.properties
             - name: DATA_PLANE_CONFIG_FILE_PATH
-              value: /etc/brokers-triggers/data
+              value: /etc/contract-resources/data
             - name: EGRESSES_INITIAL_CAPACITY
               value: "20"
             - name: INSTANCE_ID
@@ -114,11 +117,17 @@ spec:
             - name: WAIT_STARTUP_SECONDS
               value: "8"
             - name: JAVA_TOOL_OPTIONS
-              value: "-XX:+CrashOnOutOfMemoryError"
+              value: "-XX:+CrashOnOutOfMemoryError -XX:InitialRAMPercentage=70.0 -XX:MinRAMPercentage=70.0 -XX:MaxRAMPercentage=70.0"
+
           resources:
             requests:
-              cpu: 200m
-              memory: 300Mi
+              cpu: 1000m
+              # 600Mi for virtual replicas + 100Mi overhead
+              memory: 700Mi
+            limits:
+              cpu: 2000m
+              memory: 1000Mi
+
           livenessProbe:
             failureThreshold: 3
             tcpSocket:
@@ -151,12 +160,9 @@ spec:
         - name: config-kafka-broker-data-plane
           configMap:
             name: config-kafka-broker-data-plane
-        - name: kafka-broker-brokers-triggers
-          configMap:
-            name: kafka-broker-brokers-triggers
         - name: cache
           emptyDir: { }
-        - name: kafka-broker-config-logging
+        - name: kafka-config-logging
           configMap:
             name: kafka-config-logging
         - name: config-tracing
