+++ test/e2e-common.sh
@@ -215,7 +215,7 @@ function test_setup() {
 
   kubectl rollout restart statefulset -n knative-eventing kafka-source-dispatcher
   kubectl rollout restart deployment -n knative-eventing kafka-broker-receiver
-  kubectl rollout restart deployment -n knative-eventing kafka-broker-dispatcher
+  kubectl rollout restart statefulset -n knative-eventing kafka-broker-dispatcher
   kubectl rollout restart deployment -n knative-eventing kafka-sink-receiver
   kubectl rollout restart deployment -n knative-eventing kafka-channel-receiver
   kubectl rollout restart deployment -n knative-eventing kafka-channel-dispatcher
